64-bit changes -- ioctl cmds are u_longs.
SCCS-vsn: sys/netns/ns.c 8.3
SCCS-vsn: sys/netns/spp_usrreq.c 8.2
SCCS-vsn: sys/netns/ns_ip.c 8.2
SCCS-vsn: sys/netns/idp_usrreq.c 8.2

+++ usr/src/sys/netns/ns.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)ns.c	8.2 (Berkeley) %G%
+ *	@(#)ns.c	8.3 (Berkeley) %G%
  */
 
 #include <sys/param.h>
@@ -33,7 +33,7 @@ extern struct sockaddr_ns ns_netmask, ns_hostmask;
 /* ARGSUSED */
 ns_control(so, cmd, data, ifp)
 	struct socket *so;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	register struct ifnet *ifp;
 {
